# swagger.yaml
openapi: 3.0.3
info:
  title: Ibrat Lead Calling System API
  description: |
    Complete backend API for AI-powered lead calling system with admin and operator panels
    
    ## 🚀 Quick Start Guide
    
    ### **1. Authentication Flow**
    1. **Register User**: `POST /api/auth/register` (admin only)
       - Use test credentials: `john.doe@ibrat.com` / `ibrat123`
    2. **Login**: `POST /api/auth/login` with same credentials
       - Copy JWT token from response
    3. **Use APIs**: Add `Authorization: Bearer <token>` header
    
    ### **2. Test Credentials**
    - **Admin**: `john.doe@ibrat.com` / `ibrat123`
    - **Salesperson**: `jane.smith@ibrat.com` / `ibrat123`
    
    ### **3. API Access**
    - **Admin Panel**: Full system access
    - **Operator Panel**: Work interface only
    - **Real-time**: WebSocket connections
    
    ### **4. Swagger Features**
    - Click "Try it out" to test APIs
    - Use "Authorize" button for JWT tokens
    - View request/response schemas
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@ibrat.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.ibrat.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from /api/auth/login
        
        **How to use:**
        1. Register: `POST /api/auth/register`
        2. Login: `POST /api/auth/login` 
        3. Copy token from response
        4. Click "Authorize" button above
        5. Enter: `Bearer <your_token>`
        
        **Example:**
        ```
        Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjUwN2YxZjc3YmNmODZjZDc5OTQzOTAxMSIsIm5hbWUiOiJKb2huIERvZSIsImVtYWlsIjoiam9obi5kb2VAaWJyYXQuY29tIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNzM1NzE5NjAwLCJleHAiOjE3MzU3MjU2MDB9.example
        ```
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "+1234567890"
        role:
          type: string
          enum: [admin, salesperson]
          example: "salesperson"
        isActive:
          type: boolean
          example: true
        isAvailable:
          type: boolean
          example: true
        lastLogin:
          type: string
          format: date-time
        callStats:
          type: object
          properties:
            totalCalls:
              type: number
              example: 150
            successfulCalls:
              type: number
              example: 120
            totalDuration:
              type: number
              example: 7200
...
paths:
...
  /api/users:
    get:
      tags: [User Management]
      summary: Get all users
      description: Retrieve all users with pagination (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, salesperson]
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
  /api/users/available-salespeople:
    get:
      tags: [User Management]
      summary: Get available salespeople
      description: Get list of available salespeople for call assignment
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Available salespeople retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  salespeople:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
...